Extract interface for SpotifyAugmentor
